moves savedMovies state to context, moves getSavedMovies() to SavedFilms, renders saved films
